refactor: divide SidebarItem from GroupSidebar
